Change from 1 node to 3 nodes. 1 controller_manager per node.
TODO: Check the loading of the param contr_name ( and the begin of standart_arm). Sometimes it works, sometimes it does not.

The code also needs to be editted in case that the multiple node solution works . If it does work, there is no reason to have the classes being constructed with the name of the controller as a fuction param.